+++ src/liana/resource/select_resource.py
@@ -4,7 +4,7 @@ from numpy import unique
 from pandas import DataFrame, read_csv
 
 from liana._constants import DefaultValues as V
-from liana._logging import _logg
+from liana._common import _logg
 
 
 def select_resource(resource_name: str = V.resource_name) -> DataFrame: